Move TaskValidator to Utils directory
